requirements.txt
streamlit
pandas
numpy
opencv-python
ultralytics
Pillow